feat&test: Added AST support and modified tests
Added the definition of AST and modified the previous logic for obtaining SQL to generate AST.
Modified the parse method to generate SQL based on AST.
Extracted the code for preparing the parser into a separate function.
Added a function to output the AST structure, but it is not yet implemented.